# composer-service/backend/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: VoltGuard Composer API
  version: 1.0.0
  description: >
    BFF/Orchestration layer for the VoltGuard dashboard. Exposes a unified,
    token-secured REST API that orchestrates three upstream microservices:
    OAM (device registry), Anomaly Detection, and Notifications.
servers:
  - url: /api
    description: Composer base path
security:
  - bearerAuth: []
...
paths:
...
  /firmwares:
    get:
      tags: [Firmwares]
      summary: List firmware versions
      responses:
        '200':
          description: Array of Firmware objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Firmware'
              example:
                - id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  version: v1.2.0
                  uploaded_at: '2026-03-18T14:23:51.377Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'

    post:
      tags: [Firmwares]
      summary: Upload new firmware binary
      x-required-role: admin
      description: >
        Proxied to POST /firmwares on OAM. Accepts multipart/form-data,
        maximum file size 50 MB. Requires admin role.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, version]
              properties:
                file:
                  type: string
                  format: binary
                  description: Firmware binary file (max 50 MB)
                version:
                  type: string
                  description: Semantic version string, e.g. "v1.2.0"
      responses:
        '201':
          description: Firmware uploaded — returns firmware metadata from OAM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Firmware'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak access token. Role claims are read from realm_access.roles.
...
  schemas:
...
    Firmware:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: string
          description: Semantic version string, e.g. "v1.2.0"
        uploaded_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
          description: Field-level details — only present on ValidationError (400)
  responses:
    Unauthorized:
      description: Bearer token missing or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Bearer token missing or expired
    Forbidden:
      description: Authenticated but insufficient role (admin required)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: Insufficient permissions for this resource
...
    ValidationError:
      description: Validation failed — missing or invalid fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation error
            message: 'Validation failed: name is required and must be at least 2 characters'
            details:
              - name is required and must be at least 2 characters
    Conflict:
      description: Resource already exists (name or version conflict)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Conflict
            message: Resource already exists
    BadGateway:
      description: Upstream service unavailable or returned an error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Upstream service error
            message: Upstream service unavailable